feat: Use separate PDK hashes for PDK families (#731)
* Renamed `open_pdks_rev` to `pdk_hashes.yaml`
  * No change for sky130 and gf180mcu
  * Add hash for ihp-sg13g2

Enabled a range of designs for ihp-sg13g2 in the test matrix.

The define for `ihp-sg13g2` would be `PDK_ihp-sg13g2`, which is not a
valid Verilog identifier. Thus, convert all `-` to `_`.

In the metrics script I also had to split the `pdk-scl-design_name`
triple starting from the right side, since `ihp-sg13g2` contains a `-`.
(Why did they have to use a `-` in the PDK name...)

Signed-off-by: Leo Moser <leomoser99@gmail.com>