feat!: cli: reject paths with possible tilde expansion failures (#762)
This PR rejects paths provided over the command line that start with a
tilde, as it typically implies tilde expansion has failed. We considered
the alternative of implementing tilde expansion ourselves, but given
that LibreLane does not embed a shell (and that automatic tilde
expansion may in fact be surprising for some people,) we elected to
provide a *very* comprehensive error message instead:

```console
$ python3 -m librelane --pdk-root "~/.ciel"
Usage: python -m librelane [OPTIONS] [CONFIG_FILES]...
Try 'python -m librelane --help' for help.

Error: Invalid value for '--pdk-root': '~/.ciel' starts with a tilde, which is ambiguous.
  * If you meant your home directory, make sure your POSIX shell is able to expand the tilde:
    * GOOD: python -m librelane --pdk-root ~/.ciel   …
    * BAD:  python -m librelane --pdk-root=~/.ciel   …
    * BAD:  python -m librelane --pdk-root "~/.ciel" …
    * GOOD: PDK_ROOT=~/.ciel   python -m librelane …
    * BAD:  PDK_ROOT="~/.ciel" python -m librelane …
  * If you want a relative file or directory that starts with a literal "~", use an absolute path.
 ```

References:
* https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_06_01

See also:
* #725 
* https://matrix.to/#/#librelane:fossi-chat.org

---

(Additionally, I've also reformatted the changelog for style consistency, this wasn't an issue before because I was the only person writing changelogs at Efabless but now that's no longer the case.)