docs: `--prefer-upstream-nix` for installer, lix compat (#770)
Unfortunately despite the DetSys installer working best, they're
dropping upstream Nix (CppNix) support eventually, with the current flag
being needed to install upstream Nix.
https://determinate.systems/blog/installer-dropping-upstream/

Their fork of Nix is libre software, however it interacts with
presumably IPC with a proprietary daemon, `nixd`, rendering the whole
package proprietary. This is not acceptable for a free and open source
community project.

As a hedge, this also updates ciel and nix-eda to versions that lack url
literals, adding support for the [lix](https://lix.systems) community
fork of Nix.