feat!: Add SYNTH_SHOW to enable graphviz DOT file generation (#746)
This PR introduces a variable that allows to enable the `show` command
only when needed.
The `try`-`except` block around `show` did not prevent LibreLane from
stopping when the design is empty. However, this is the case when
generating DEF templates for e.g. Tiny Tapeout or HeiChips.

The default behavior has been changed to `default=False`, as most user
won't need it and it may slow down large designs and take up disk space.

Signed-off-by: Leo Moser <leomoser99@gmail.com>