feat: Yosys.Synthesis: Added variables to set `keep_hierarchy` (#760)
* `SYNTH_KEEP_HIERARCHY_MIN_COST`: Sets the `keep_hierarchy` attribute
on modules where the gate count is estimated to exceed the specified
threshold. This prevents larger modules from being flattened.
* `SYNTH_KEEP_HIERARCHY_INSTANCES`: A list of instances for which to set
the `keep_hierarchy` attribute.
* `SYNTH_KEEP_HIERARCHY_MODULES`: A list of modules for which to set the
`keep_hierarchy` attribute.

Two things are important to note:

- In `check_macro_instances.tcl` I had to remove `-hierarchical` on
`get_cells` since this would search for the specified cell in each
hierarchy level. This works fine when everything is flattened, but
doesn't when you have a hierarchy and specify the hierarchical path to a
cell, i.e., `a/b/c`.
- I had to set the `keep_hierarchy` attribute on all blackbox modules,
since `keep_hierarchy -min_cost` would otherwise simply stop when
encountering a bb module. (The other option would be to set a min_cost.)

---------

Signed-off-by: Leo Moser <leomoser99@gmail.com>