chore!: bump python requirement to 3.10
For a broad audience: this drops support for Ubuntu versions below 22.04 and Debian versions below 12. It also requires an up-to-date version of Python to be installed for macOS.

On our side, this unlocks poetry 2.0.0, which in turn unlocks the [PEP-0621](https://peps.python.org/pep-0621) format for `pyproject.toml`. But it also unlocks a neat bunch of language features we'd not been able to use up to this point:

- https://docs.python.org/3/whatsnew/3.9.html
- https://docs.python.org/3/whatsnew/3.10.html

This commit also updates mypy and does a number of small fixes to ensure that it still clears this commit. Additionally, the CI has been configured to also to test Python 3.13 and the brand-new 3.14.