ci: fix contributor CI not using cache (#799)
We've noticed that our external contributors are not getting `vars.`
which causes their builds to take forever and fail.

We now always add the FOSSi Foundation Cache as a backup. The OpenLane
cachix has been removed.

This also adds some minor fixes to both the GitHub Actions and the
ReadTheDocs CI, as well as tightening up mutual exclusion for the CI
such that:
- pushes to feature branches and pull requests pre-empt previous CI runs
- pushes to main, dev and version branches are serialized