Fix subprocess formatting (#817)
This PR fixes a long outstanding issue that has to do with the
formatting of subprocess output.

Initially, I thought the change would be needed in the logger or Rich
itself, but it turns out that the subprocess output was completely
stripped of its whitespace ^^

This changes the call from `strip` to `rstrip`, as it still allows Rich
to do a better formatting of the output if there is whitespace after the
text, while keeping the formatting of tables and such.